backend_sgos_py/requirements.txt
fastapi==0.109.0
uvicorn==0.24.0
sqlalchemy==2.0.43
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.11.5
pydantic-settings==2.1.0
fastapi-mail==1.4.1
jinja2==3.1.2
pytz==2023.3
email-validator==2.2.0
